+++ Patches/EntranceTeleportPatch.cs
@@ -1,5 +1,4 @@
 ﻿using GameNetcodeStuff;
-using GeneralImprovements.OtherMods;
 using GeneralImprovements.Utilities;
 using HarmonyLib;
 using System.Reflection;
@@ -31,7 +30,7 @@ namespace GeneralImprovements.Patches
             // If configured, add a scan node to all exits and fire entrances
             if (Plugin.ShowDoorsOnScanner.Value && !(__instance.entranceId == 0 && __instance.isEntranceToBuilding))
             {
-                string text = __instance.isEntranceToBuilding ? "Fire Entrance" : __instance.entranceId == 0 ? "Main Exit" : $"Fire Exit{(MimicsHelper.IsActive ? "?" : "")}";
+                string text = __instance.isEntranceToBuilding ? "Fire Entrance" : __instance.entranceId == 0 ? "Main Exit" : $"Fire Exit{(OtherModHelper.MimicsActive ? "?" : "")}";
                 ObjectHelper.CreateScanNodeOnObject(__instance.gameObject, 0, 1, __instance.isEntranceToBuilding ? 50 : 20, text);
             }
         }